requirements.txt
# 基础依赖
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=9.0.0

# 深度学习框架
torch>=1.10.0
torchvision>=0.11.0

# 数据处理
pandas>=1.3.0
scikit-learn>=0.24.0

# 网络请求
requests>=2.26.0
aiohttp>=3.8.0

# 图像处理
albumentations>=1.1.0
scipy>=1.7.0

# 进度条和日志
tqdm>=4.62.0
tensorboard>=2.7.0
loguru>=0.5.3

# 配置管理
pyyaml>=5.4.0
python-dotenv>=0.19.0

# 测试
pytest>=6.2.0
pytest-cov>=2.12.0

# 代码质量
black>=21.7b0
flake8>=3.9.0
isort>=5.9.0